separate dev app from library code + useTimer, useStopwatch and useTime to different files
